test boilerplate

+++ cunumeric/deferred.py
@@ -1082,9 +1082,7 @@ class DeferredArray(NumPyThunk):
                 view.copy(rhs, deep=False)
 
     def broadcast_to(self, shape: NdShape) -> NumPyThunk:
-        return DeferredArray(
-            self.runtime, base=self._broadcast(shape), dtype=self.dtype
-        )
+        return DeferredArray(self.runtime, base=self._broadcast(shape))
 
     def reshape(self, newshape: NdShape, order: OrderType) -> NumPyThunk:
         assert isinstance(newshape, Iterable)
